WIP extract denormalization method
